refactor: decompose evaluation/evaluator.py into focused sub-modules
